requirements.txt
blinker==1.9.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
faiss-cpu==1.12.0
filelock==3.19.1
Flask==3.1.2
fsspec==2025.9.0
huggingface-hub==0.35.3
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.2
lxml==6.0.2
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.5
numpy==2.3.3
packaging==25.0
pillow==11.3.0
PyPDF2==3.0.1
python-docx==1.2.0
PyYAML==6.0.3
regex==2025.9.18
requests==2.32.5
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.2
sentence-transformers==5.1.1
setuptools==80.9.0
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.22.1
torch==2.8.0
tqdm==4.67.1
transformers==4.56.2
typing_extensions==4.15.0
urllib3==2.5.0
Werkzeug==3.1.3
